+++ Assets/Plugins/Appboy/BrazePlatform.cs
@@ -106,4 +106,7 @@ public interface BrazePlatform {
   void DisplayContentCards();
   void AddToSubscriptionGroup(string id);
   void RemoveFromSubscriptionGroup(string id);
+  void RefreshFeatureFlags();
+  FeatureFlag GetFeatureFlag(string id);
+  List<FeatureFlag> GetAllFeatureFlags();
 }