+++ unity-samples/Assets/Scripts/UnitySampleConstants.cs
@@ -51,6 +51,7 @@ namespace Utilities {
     public static string LogPurchaseScene = "LogPurchase";
     public static string SetUserFieldScene = "SetUserField";
     public static string CustomAttributesSettingScene = "CustomAttributesSetting";
+    public static string FeatureFlagScene = "FeatureFlags";
     
     // Change User
     public static string ChangeUserCancelLog = "Cancel changing User ID.";